+++ lib/features/achievement_screen/presentation/blocs/save_weight_bloc/save_weight_event.dart
@@ -1,7 +1,7 @@
 part of 'save_weight_bloc.dart';
 
  class SaveWeightEvent extends Equatable {
-  final int weight;
+  final double weight;
   const SaveWeightEvent({required this.weight});
 
   @override
